backend/requirements.txt
channels==4.0.0
daphne==4.0.0
Django==4.1.7
django-debug-toolbar==4.1.0
django-cors-headers==3.14.0
django-polymorphic==3.1.0
djangorestframework==3.14.0
djoser==2.2.0
Pillow==9.4.0
psycopg2-binary==2.9.5
